Mobile-test/requirements.txt
Appium-Python-Client==2.2.4
pytest
allure-pytest
selenium